+++ indico/modules/events/client/js/util/list_generator.js
@@ -36,7 +36,22 @@
   }
 
   global.handleRowSelection = function(trigger) {
-    const $obj = $('table.i-table input.select-row').on('change', function() {
+    let lastClickedIndex;
+    const $obj = $('table.i-table input.select-row');
+    $obj.on('click', function(e) {
+      const currentIndex = $obj.index(this);
+      if (e.shiftKey) {
+        const start = Math.min(lastClickedIndex, currentIndex);
+        const end = Math.max(lastClickedIndex, currentIndex);
+        $obj
+          .slice(start, end + 1)
+          .prop('checked', this.checked)
+          .trigger('change');
+      }
+      lastClickedIndex = currentIndex;
+    });
+
+    $obj.on('change', function() {
       $(this).closest('tr').toggleClass('selected', this.checked);
       $('.js-requires-selected-row').toggleClass(
         'disabled',